+++ packages/rsi-interpretation-player/streamers/index.js
@@ -1,2 +1,4 @@
-import Akkadu from '@akkadu/rtc-streamer-consumer' 
+import Akkadu from '@akkadu/rtc-streamer-consumer'
+import '../src/index.scss'
+
 export default Akkadu