+++ packages/rsi-react-example/src/components/interpretation-player.jsx
@@ -32,8 +32,6 @@ class InterpretationPlayer extends Component {
         <h3> Interpretation Player: </h3>
        <RsiApiInterpretationPlayer
            sdkKey="d9ed0b1e-6027-4c19-a43a-69b231713f1c"
-           positionMenu="bottom" 
-           isBoxShadow={false}
            onReady = {handleOnReady}
            onLanguageSelected = {handleOnLanguageSelected}
            onConnectionStatusUpdated = {handleOnConnectionStatusUpdated}