+++ packages/rsi-react/example/src/App.js
@@ -1,6 +1,4 @@
-
 import React from 'react'
-
 import { RsiApiInterpretationPlayer } from '@akkadu/rsi-react'
 import { RsiInterpretationManager } from '@akkadu/rsi-react'
 import '@akkadu/rsi-react/dist/index.css'
@@ -29,7 +27,8 @@ const handleOnCreateEvent = (e) => {
   console.info('receive onConnectionStatusUpdated event:',e );
 }
 const App = () => {
-  return  <div>
+  return  <div style={{ marginBottom: '2rem' }}>
+    <h3>Interpretation Player</h3>
     <video loop="" controls width="640" height="480">
       <source type="video/mp4" src="https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4"/>
     </video>
@@ -41,6 +40,7 @@ const App = () => {
         isBoxShadow={false}
         isPlayerControlled={true}
    />
+   <h3 style={{ margin: '2rem 0 0.5rem 0' }}>Interpretation Manager</h3>
     <RsiInterpretationManager
         sdkKey="d9ed0b1e-6027-4c19-a43a-69b231713f1c"
         onCreateEvent = {handleOnCreateEvent}