+++ packages/rsi-vanilla/update-version.js
@@ -19,6 +19,7 @@ const currentVersion = packageJson.version
 const myBucket = 'akkadu-assets'
 const filePathInterpretationPlayer = path.resolve('./packages/rsi-vanilla/dist/interpretation-player.min.js')
 const filePathInterpretationManager = path.resolve('./packages/rsi-vanilla/dist/interpretation-manager.min.js')
+const filePathInterpretationPlayerCss = path.resolve('./packages/rsi-vanilla/dist/index.min.css')
 
 let tmpVersionFile = require(path.resolve('./packages/rsi-vanilla/tmp-version.json'));
 let tmpVersion;
@@ -49,9 +50,9 @@ if(tmpVersionFile && tmpVersionFile.tmpVersion){
 /**
  * @description To update the vanillaJs Version documentation with the new version.
  */
- const concatMarkdownDoc = (fileName) =>{
+ const concatMarkdownDoc = (fileName, ext) =>{
   fs.readFile('./rsi-documentation/docs/vanilla-js/versions.md', 'utf8', (err, doc) => {
-    const newVersion = `- Version ${currentVersion}: https://akkadu-assets.s3.amazonaws.com/akkadu-rsi-widget/rsi-vanilla/${currentVersion}/${fileName}.min.js`
+    const newVersion = `- Version ${currentVersion}: https://akkadu-assets.s3.amazonaws.com/akkadu-rsi-widget/rsi-vanilla/${currentVersion}/${fileName}.min.${ext}`
     const newDoc = `${doc} \n ${newVersion}`
     fs.writeFile(`./rsi-documentation/docs/vanilla-js/versions.md`, newDoc, function(err, result) {
       if(err) console.error('error when updating the doc with the new version', err);
@@ -65,20 +66,21 @@ console.info('Updating the version on our bucket akkadu-rsi-widget/rsi-vanilla/,
 console.info('🚧🚧🚧');
 console.info('currentVersion',currentVersion);
 
-const updateAWS = (file, fileName) =>{
+const updateAWS = (file, fileName, ext, contentType = 'application/javascript') =>{
   fs.readFile(file, function (err, data) {
     if (err) { throw err; }
     var base64data = new Buffer(data, 'binary');
     var s3 = new AWS.S3();
     s3.putObject({
       Bucket: myBucket,
-      Key: `akkadu-rsi-widget/rsi-vanilla/${currentVersion}/${fileName}.min.js`,
+      Key: `akkadu-rsi-widget/rsi-vanilla/${currentVersion}/${fileName}.min.${ext}`,
       Body: base64data,
-      ACL: 'public-read'
+      ACL: 'public-read',
+      ContentType: contentType
     },function (resp) {
       console.info(`Successfully uploaded the new version of rsi-widget on AWS, version: ${currentVersion}.`);
       writeTmpVersion();
-      concatMarkdownDoc(fileName)
+      concatMarkdownDoc(fileName, ext)
     });
 });
 }
@@ -87,8 +89,9 @@ const updateAWS = (file, fileName) =>{
 AWS.config.update({ accessKeyId: process.env.AWS_ACCESS_KEY, secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY, });
 
 
-updateAWS(filePathInterpretationPlayer, 'interpretation-player')
-updateAWS(filePathInterpretationManager, 'interpretation-manager')
+updateAWS(filePathInterpretationPlayer, 'interpretation-player', 'js')
+updateAWS(filePathInterpretationPlayerCss, 'interpretation-player', 'css', 'text/css')
+updateAWS(filePathInterpretationManager, 'interpretation-manager', 'js')