+++ packages/rsi-vue/dev/serve.vue
@@ -42,9 +42,7 @@ export default Vue.extend({
     <br>
    <h3>Interpretation Player:</h3>
     <rsi-api-interpretation-player 
-      sdkKey="d9ed0b1e-6027-4c19-a43a-69b231713f1c" 
-      positionMenu="bottom" 
-      :isBoxShadow="true"
+      sdkKey="d9ed0b1e-6027-4c19-a43a-69b231713f1c"
       :isPlayerControlled="true"
       v-on:onLanguageSelected="handleOnLanguageSelected"
       v-on:onReady="handleOnReady"