+++ packages/rsi-interpretation-manager/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-manager@0.1.3...@akkadu/rsi-interpretation-manager@0.1.4) (2021-07-08)
+
+**Note:** Version bump only for package @akkadu/rsi-interpretation-manager
+
+
+
+
+
+## [0.1.3](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-manager@0.1.2...@akkadu/rsi-interpretation-manager@0.1.3) (2021-07-08)
+
+**Note:** Version bump only for package @akkadu/rsi-interpretation-manager
+
+
+
+
+
+## [0.1.2](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-manager@0.1.1...@akkadu/rsi-interpretation-manager@0.1.2) (2021-07-08)
+
+**Note:** Version bump only for package @akkadu/rsi-interpretation-manager
+
+
+
+
+
 ## [0.1.1](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-manager@0.1.0...@akkadu/rsi-interpretation-manager@0.1.1) (2021-07-02)
 
 **Note:** Version bump only for package @akkadu/rsi-interpretation-manager