+++ packages/rsi-react-example/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.24](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-react-example@0.1.23...@akkadu/rsi-react-example@0.1.24) (2021-07-12)
+
+**Note:** Version bump only for package @akkadu/rsi-react-example
+
+
+
+
+
+## [0.1.23](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-react-example@0.1.22...@akkadu/rsi-react-example@0.1.23) (2021-07-08)
+
+**Note:** Version bump only for package @akkadu/rsi-react-example
+
+
+
+
+
+## [0.1.22](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-react-example@0.1.21...@akkadu/rsi-react-example@0.1.22) (2021-07-08)
+
+**Note:** Version bump only for package @akkadu/rsi-react-example
+
+
+
+
+
+## [0.1.21](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-react-example@0.1.20...@akkadu/rsi-react-example@0.1.21) (2021-07-08)
+
+**Note:** Version bump only for package @akkadu/rsi-react-example
+
+
+
+
+
 ## [0.1.20](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-react-example@0.1.19...@akkadu/rsi-react-example@0.1.20) (2021-07-02)
 
 **Note:** Version bump only for package @akkadu/rsi-react-example