+++ packages/rsi-vanilla/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.14](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.13...@akkadu/rsi-vanilla@1.0.14) (2021-07-12)
+
+**Note:** Version bump only for package @akkadu/rsi-vanilla
+
+
+
+
+
+## [1.0.13](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.12...@akkadu/rsi-vanilla@1.0.13) (2021-07-08)
+
+**Note:** Version bump only for package @akkadu/rsi-vanilla
+
+
+
+
+
+## [1.0.12](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.11...@akkadu/rsi-vanilla@1.0.12) (2021-07-08)
+
+**Note:** Version bump only for package @akkadu/rsi-vanilla
+
+
+
+
+
+## [1.0.11](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.10...@akkadu/rsi-vanilla@1.0.11) (2021-07-08)
+
+**Note:** Version bump only for package @akkadu/rsi-vanilla
+
+
+
+
+
 ## [1.0.10](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.9...@akkadu/rsi-vanilla@1.0.10) (2021-07-02)
 
 **Note:** Version bump only for package @akkadu/rsi-vanilla